Creating a base command class.
