--- src/Robo/BltRobo.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Acquia\Blt\Robo;
-
-use Robo\Robo;
-
-class BltRobo extends Robo {
-
-  /**
-   * Override parent createDefaultApplication() to use custom application.
-   */
-  public static function createDefaultApplication($appName = null, $appVersion = null)
-  {
-    $appName = $appName ?: self::APPLICATION_NAME;
-    $appVersion = $appVersion ?: self::VERSION;
-
-    $app = new BltApplication($appName, $appVersion);
-    $app->setAutoExit(false);
-    return $app;
-  }
-
-}
