+++ src/Robo/Command/Test.php
@@ -1,59 +1,16 @@
 <?php
 
 namespace Acquia\Blt\Robo\Command;
-use Symfony\Component\Yaml\Yaml;
+
+use Acquia\Blt\Robo\Command\BltTasks;
 
 /**
  * This is project's console commands configuration for Robo task runner.
  *
  * @see http://robo.li/
  */
-class RoboFile extends \Robo\Tasks
+class Test extends BltTasks
 {
-
-  protected $config = [];
-  protected $repoRoot;
-  protected $bltRoot;
-  protected $docroot;
-  protected $bin;
-
-  public function __construct() {
-    $this->setRepoRoot();
-    $this->setBltRoot();
-    $this->docroot = "{$this->repoRoot}/docroot";
-    $this->setConfig();
-    $this->bin = "{$this->repoRoot}/vendor/bin";
-  }
-
-  public function setRepoRoot() {
-    $possible_repo_roots = [
-      $_SERVER['PWD'],
-      getcwd(),
-    ];
-
-    foreach ($possible_repo_roots as $possible_repo_root) {
-      if (file_exists("$possible_repo_root/blt/project.yml")) {
-        $this->repoRoot = $possible_repo_root;
-        break;
-      }
-    }
-  }
-
-  public function setBltRoot() {
-    $this->bltRoot = dirname(dirname(dirname(dirname(__FILE__))));
-  }
-
-  public function setConfig() {
-    $default_config =  Yaml::parse(file_get_contents("{$this->bltRoot}/phing/build.yml"));
-    $this->config = Yaml::parse(file_get_contents("{$this->repoRoot}/blt/project.yml"));
-    $this->config = $this->array_merge_recursive_distinct($this->config, $default_config);
-
-    array_walk_recursive($this->config, function (&$value, $key) {
-      $value = str_replace('${repo.root}', $this->repoRoot, $value);
-      $value = str_replace('${blt.root}', $this->bltRoot, $value);
-      $value = str_replace('${docroot}', $this->docroot, $value);
-    });
-  }
 
   /**
    * Runs all tests, including Behat, PHPUnit, and Security Update check.
@@ -78,7 +35,6 @@ class RoboFile extends \Robo\Tasks
    * @command tests:behat
    */
   public function testsBehat() {
-    $this->yell("Repo root is {$this->repoRoot}");
     $this->killByPort('4444');
     $this->killByName('selenium');
     $this->killByName('phantomjs');
@@ -101,6 +57,9 @@ class RoboFile extends \Robo\Tasks
 
   }
 
+  /**
+   *
+   */
   protected function launchPhantomJs() {
     $this->verifyPhantomJsConfig();
 
@@ -113,6 +72,9 @@ class RoboFile extends \Robo\Tasks
     // Wait for http://127.0.0.1:4444/wd/hub to become available.
   }
 
+  /**
+   * @throws \Exception
+   */
   protected function verifyPhantomJsConfig() {
     $result = $this->_exec("grep 'jakoch/phantomjs-installer' composer.json");
     if (!$result->wasSuccessful()) {
@@ -143,36 +105,20 @@ class RoboFile extends \Robo\Tasks
     }
   }
 
+  /**
+   * @param $port
+   */
   protected function killByPort($port) {
     $this->say("Killing all processes on port $port");
     $this->_exec("lsof -ti tcp:$port | xargs kill");
   }
 
+  /**
+   * @param $name
+   */
   protected function killByName($name) {
     $this->say("Killing all processing containing string '$name'");
     $this->_exec("pgrep $name | xargs kill");
   }
 
-  protected function killPhantomJs() {
-
-  }
-
-  protected function array_merge_recursive_distinct ( array &$array1, array &$array2 )
-  {
-    $merged = $array1;
-
-    foreach ( $array2 as $key => &$value )
-    {
-      if ( is_array ( $value ) && isset ( $merged [$key] ) && is_array ( $merged [$key] ) )
-      {
-        $merged [$key] = $this->array_merge_recursive_distinct ( $merged [$key], $value );
-      }
-      else
-      {
-        $merged [$key] = $value;
-      }
-    }
-
-    return $merged;
-  }
 }
